src/ai_o11y_custos/scoring_function/__init__.py
import datetime
import logging
import azure.functions as func

def main(mytimer: func.TimerRequest) -> None:
    utc_timestamp = datetime.datetime.utcnow().replace(tzinfo=datetime.timezone.utc).isoformat()
    logging.info("AI_O11y_Custos scoring tick at %s", utc_timestamp)
    # TODO: ler custos do ADLS, calcular probabilidade e escrever em Log Analytics/Storage.
