requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
lightgbm==4.5.0
statsmodels==0.14.2
mlflow==2.15.1
# Azure SDKs (ajuste conforme uso real)
azure-identity==1.17.1
azure-storage-blob==12.19.1
azure-monitor-query==1.5.2
# Funções
azure-functions==1.20.0
